tests: fold small substrate tests into harness
